backend/pyproject.toml
[project]
name = "mmattend-backend"
version = "0.1.0"
description = "Production FastAPI backend for the AI attendance platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.8",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.38",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.10.6",
    "pydantic-settings>=2.7.1",
    "supabase>=2.13.0",
    "httpx>=0.28.1",
    "PyJWT[crypto]>=2.10.1",
    "pgvector>=0.3.6",
    "insightface>=0.7.3",
    "onnxruntime>=1.17.0",
    "opencv-python-headless>=4.9.0",
    "numpy>=1.26.0",
    "slowapi>=0.1.9",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.7",
]

[build-system]
requires = ["setuptools>=75.8.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]